Changes asked by lecturer and some code improvement
Changed Tree output: specified where tree element is GOTO label and
where it is GOTO operand, wrote IF and ELSE parts of conditional
statement;
Removed unnecessary type conversion

+++ Compiler_SI/Program.cs
@@ -114,7 +114,7 @@ namespace Compiler_SI
                 switch (token.Type)
                 {
                     case 1:
-                        id = (int) sc.GetTokenValue(token).ToCharArray()[0];
+                        id = sc.GetTokenValue(token).ToCharArray()[0];
                         break;
                     case 0:
                         id = 401 + token.Index;
